+++ src/hikari_cp/core.clj
@@ -11,7 +11,8 @@
    :idle-timeout       600000
    :max-lifetime       1800000
    :minimum-idle       10
-   :maximum-pool-size  10})
+   :maximum-pool-size  10
+   :register-mbeans    false})
 
 (def ^{:private true} adapters-to-datasource-class-names
   {"derby"          "org.apache.derby.jdbc.ClientDataSource"
@@ -76,6 +77,7 @@
    :maximum-pool-size  IntGte1
    :adapter            AdaptersList
    (s/optional-key :leak-detection-threshold) IntGte2000
+   :register-mbeans    s/Bool
    s/Keyword           s/Any})
 
 (defn- exception-message
@@ -120,7 +122,8 @@
                 pool-name
                 read-only
                 username
-                leak-detection-threshold]} options
+                leak-detection-threshold
+                register-mbeans]} options
         datasource-class-name (get
                                adapters-to-datasource-class-names
                                adapter)]
@@ -134,7 +137,8 @@
       (.setMaxLifetime         max-lifetime)
       (.setMinimumIdle         minimum-idle)
       (.setMaximumPoolSize     maximum-pool-size)
-      (.setDataSourceClassName datasource-class-name))
+      (.setDataSourceClassName datasource-class-name)
+      (.setRegisterMbeans      register-mbeans))
 
     ;; Set optional properties
     (if username (.setUsername config username))